Moving extensions file to domain package.
