+++ app/src/main/java/br/stone/mobiletraining/samilasantos/ex02cadastrodocliente/data/common/Extensions.kt
@@ -1,4 +1,4 @@
-package br.stone.mobiletraining.samilasantos.ex02cadastrodocliente.ui.common
+package br.stone.mobiletraining.samilasantos.ex02cadastrodocliente.data.common
 
 import android.app.Activity
 import android.app.AlertDialog
