+++ app/src/main/java/br/stone/mobiletraining/samilasantos/ex02cadastrodocliente/ui/registerScreen/RegisterViewModel.kt
@@ -1,11 +1,11 @@
 package br.stone.mobiletraining.samilasantos.ex02cadastrodocliente.ui.registerScreen
 
+import br.stone.mobiletraining.samilasantos.ex02cadastrodocliente.data.common.toDate
 import br.stone.mobiletraining.samilasantos.ex02cadastrodocliente.domain.Entrepreneur
 import br.stone.mobiletraining.samilasantos.ex02cadastrodocliente.domain.Repository
 import br.stone.mobiletraining.samilasantos.ex02cadastrodocliente.domain.common.Result
 import br.stone.mobiletraining.samilasantos.ex02cadastrodocliente.domain.uc.CreateEntrepreneur
 import br.stone.mobiletraining.samilasantos.ex02cadastrodocliente.ui.common.ErrorCode
-import br.stone.mobiletraining.samilasantos.ex02cadastrodocliente.ui.common.toDate
 
 class RegisterViewModel(private val repository: Repository) {
